test: provide typedefs for device management and not found routes
Although they're typed being concatenated with the app-wide routes in
"src/Routes.ts", it's quite difficult to inspect the type mismatch error
messages there since an incorrect route isn't highlighted. So, it seems
a good idea to provide typedef layer earlier, before concatenation.
