+++ src/routing/routes/not-found.ts
@@ -1,6 +1,9 @@
+import { Route } from '../typings';
+
+// components
 import { NoMatch } from '../../pages';
 
-export const routes = [
+export const routes: Route[] = [
   {
     main: NoMatch,
     label: 'Not found',
