+++ src/routing/typings.ts
@@ -1,0 +1,10 @@
+import { ComponentType } from 'react';
+
+export type Route = {
+  path?: string;
+  label: string;
+  icon?: string;
+  main?: ComponentType;
+  exact?: boolean;
+  routes?: Route[];
+};
